+++ main.go
@@ -4,111 +4,22 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
-	"time"
+
+	p "github.com/Yandex-Practicum/go-db-sql-final/parcel"
 
 	_ "modernc.org/sqlite"
 )
 
-const (
-	ParcelStatusRegistered = "registered"
-	ParcelStatusSent       = "sent"
-	ParcelStatusDelivered  = "delivered"
-	PathDB                 = "tracker.db"
-	DBDriver               = "sqlite"
-)
-
-type Parcel struct {
-	Number    int    //номер посылки
-	Client    int    //идентификатор клиента
-	Status    string //статус посылки
-	Address   string //адрес посылки
-	CreatedAt string //дата и время создания посылки
-}
-
-type ParcelService struct {
-	store ParcelStore
-}
-
-func NewParcelService(store ParcelStore) ParcelService {
-	return ParcelService{store: store}
-}
-
-func (s ParcelService) Register(client int, address string) (Parcel, error) {
-	parcel := Parcel{
-		Client:    client,
-		Status:    ParcelStatusRegistered,
-		Address:   address,
-		CreatedAt: time.Now().UTC().Format(time.RFC3339),
-	}
-
-	id, err := s.store.Add(parcel)
-	if err != nil {
-		return parcel, err
-	}
-
-	parcel.Number = id
-
-	fmt.Printf("Новая посылка № %d на адрес %s от клиента с идентификатором %d зарегистрирована %s\n",
-		parcel.Number, parcel.Address, parcel.Client, parcel.CreatedAt)
-
-	return parcel, nil
-}
-
-func (s ParcelService) PrintClientParcels(client int) error {
-	parcels, err := s.store.GetByClient(client)
-	if err != nil {
-		return err
-	}
-
-	fmt.Printf("Посылки клиента %d:\n", client)
-	for _, parcel := range parcels {
-		fmt.Printf("Посылка № %d на адрес %s от клиента с идентификатором %d зарегистрирована %s, статус %s\n",
-			parcel.Number, parcel.Address, parcel.Client, parcel.CreatedAt, parcel.Status)
-	}
-	fmt.Println()
-
-	return nil
-}
-
-func (s ParcelService) NextStatus(number int) error {
-	parcel, err := s.store.Get(number)
-	if err != nil {
-		return err
-	}
-
-	var nextStatus string
-	switch parcel.Status {
-	case ParcelStatusRegistered:
-		nextStatus = ParcelStatusSent
-	case ParcelStatusSent:
-		nextStatus = ParcelStatusDelivered
-	case ParcelStatusDelivered:
-		return nil
-	}
-
-	fmt.Printf("У посылки № %d новый статус: %s\n", number, nextStatus)
-
-	return s.store.SetStatus(number, nextStatus)
-}
-
-func (s ParcelService) ChangeAddress(number int, address string) error {
-	return s.store.SetAddress(number, address)
-}
-
-func (s ParcelService) Delete(number int) error {
-	return s.store.Delete(number)
-}
-
 func main() {
-	db, err := sql.Open(DBDriver, PathDB)
+	db, err := sql.Open(p.DBDriver, p.PathDB)
 	if err != nil {
 		log.Println("can't open DB:", err)
 	}
 	defer db.Close()
 
-	store := NewParcelStore(db)
+	store := p.NewParcelStore(db)
 
-	service := NewParcelService(store)
+	service := p.NewParcelService(store)
 
 	// регистрация посылки
 	client := 1
