+++ parcel/parcel.go
@@ -1,9 +1,22 @@
-package main
+package parcel
 
 import (
 	"database/sql"
 	"log"
 )
+
+const (
+	PathDB   = "db/tracker.db"
+	DBDriver = "sqlite"
+)
+
+type Parcel struct {
+	Number    int    //номер посылки
+	Client    int    //идентификатор клиента
+	Status    string //статус посылки
+	Address   string //адрес посылки
+	CreatedAt string //дата и время создания посылки
+}
 
 type ParcelStore struct {
 	db *sql.DB
